Restructure runtime crate
