+++ runtime/Cargo.toml
@@ -106,7 +106,7 @@ zrml-swaps-runtime-api = { default-features = false, path = "../zrml/swaps/runti
 sp-io = { branch = "moonbeam-polkadot-v0.9.19", git = "https://github.com/purestake/substrate" }
 
 [features]
-default = ["std"]
+default = ["std", "runtime-zeitgeist"]
 parachain = [
     # Cumulus
 
@@ -140,7 +140,7 @@ parachain = [
 
     "log"
 ]
-
+runtime-battery-station=[]
 runtime-benchmarks = [
     "cumulus-pallet-xcmp-queue/runtime-benchmarks",
     "frame-benchmarking/runtime-benchmarks",
@@ -178,6 +178,7 @@ runtime-benchmarks = [
     "zrml-simple-disputes/runtime-benchmarks",
     "zrml-swaps/runtime-benchmarks",
 ]
+runtime-zeitgeist=[]
 std = [
     "frame-executive/std",
     "frame-support/std",
