+++ runtime/src/battery_station/mod.rs
@@ -1,0 +1,1 @@
+#![cfg(feature = "runtime-battery-station")]
