+++ runtime/src/zeitgeist/mod.rs
@@ -1,0 +1,1 @@
+#![cfg(feature = "runtime-zeitgeist")]
